Update LICENSE
                          💖**Mycrypto.com-BETA/LICENSE**💖
                                      *****************************
      

https://github.com/https-github-com-52005723-ruzyysmartt/MyCrypto.com-BETA.git  This is licensed under the :>

⚖ MIT LICENSE
**************
           

A short and simple permissive license with conditions only requiring preservation of copyright and license notices. 

Licensed works, modifications, and larger works may be distributed under different terms and without source code.

## MIT License

Copyright (c) 2019-2020 ** Mycrypto.com-BETA LLC

Copyright (c) 2020 MyCrypto.com Inc.

Permission is hereby granted, free of charge, to any person obtaining a copy

of this software and associated documentation files (the "Software"), to deal

in the Software without restriction, including without limitation the rights

to use, copy, modify, merge, publish, distribute, sublicense, and/or sell

copies of the Software, and to permit persons to whom the Software is

furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in all

copies or substantial portions of the Software.

THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR

IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,

FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE

AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER

LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,

OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE

SOFTWARE.

more info:
MyCrypto.com-BETA /LICENSE
https://github.com/https-github-com-52005723-ruzyysmartt/MyCrypto.com-BETA.git
MIT License
Ashort and simple permissive license with conditions only requiring preservation of copyright and license notices. 

Adapted from: https://github.com/ruzyysmartt/Mycrypto.com-BETA

Adapted from:https://github.com/ruzyysmartt/mycrypto.com-BETA

@ GitHub.2020